Use a stable version of commons-imaging
It seems that Apache Commons Imaging has had some past releases...
but not accessible from a public Maven repository that I could find.

However, I was able to build the source at the tag IMAGING_1_0_RC7
(which claims to be version 1.0 in its POM), and deploy it to the
ImageJ Nexus's 3rdparty repository. And fortunately, that version
is compile-time compatible with this project.